ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
numpy==2.2.1
Pillow==11.1.0
pyzbar==0.1.9
rapidfuzz==3.11.0
ImageHash==4.3.1
pdf2image==1.16.3
pymongo==4.9.2
python-dotenv==1.0.1
python-slugify==8.0.1
easyocr==1.6.3

